update entrypoint.sh in k8s
We build clickhouse cluster in k8s by clickhouse-operator. 
clickhouse-server.log is mounted by emptydir on the path `/var/log/clickhouse-server`.
It causes the file system error called `Necessary directory '/var/log/clickhouse-server' isn't owned by user with id '101' `  when  CLICKHOUSE_DO_NOT_CHOWN is set to true(1).
Because emptydir file and dir is owned by root when it started, it has to do the chown cmd.
But it will take a long time to run chown cmd when the cluster has much data (in table data dir), the cluster cannot even restart because of  `initialDelaySeconds`
So chown operation is necessary to check if the file owner and group is owned by user 101.